backend: rename sync/ to refresh/ which is how we talk about it elsewhere
